refactor(engine): fold wardian-engine into wardian-core::engine
