+++ crates/wardian-core/Cargo.toml
@@ -8,6 +8,7 @@ repository.workspace = true
 homepage.workspace = true
 
 [dependencies]
+async-trait = "0.1"
 chrono = { workspace = true }
 dirs = { workspace = true }
 once_cell = { workspace = true }
@@ -16,6 +17,7 @@ serde = { workspace = true }
 serde_json = { workspace = true, features = ["preserve_order"] }
 serde_norway = "0.9"
 thiserror = { workspace = true }
+tokio = { version = "1.49.0", features = ["sync", "macros", "rt", "rt-multi-thread", "time"] }
 uuid = { workspace = true }
 
 [dev-dependencies]
