+++ crates/wardian-core/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod control;
 pub mod db;
+pub mod engine;
 pub mod identity;
 pub mod models;
 pub mod paths;
